SimpleArgParse.jl Documentation

SimpleArgParse.jl Documentation

Types

Functions

Constants

Index

Types

SimpleArgParse.AbstractValidator — Type

AbstractValidator

The supertype for validators.

SimpleArgParse.ArgForms — Type

ArgForms

Command-line arguments, short and long forms.

Fields

short::String

long::String

SimpleArgParse.ArgumentParser — Type

ArgumentParser

Command-line argument parser with numkey-value stores and attributes.

Fields

stores

kv_store::OrderedDict{UInt16,ArgumentValues} = OrderedDict(): numkey => value store: { numkey: ArgumentValues(value, type, required, help) }; up to 65,536 argument keys

arg_store::OrderedDict{String,UInt16} = OrderedDict(): numkey-value store: arg => numkey

lng::UInt16 = 0: counter of stored args

attributes

filename::String = "": file name

description::String = "": description

authors::Vector{String} = String[]: name of author(s): First Last <first.last@email.address>

documentation::String = "": URL of documentations

repository::String = "": URL of software repository

license::String = "": name of license

usage::String = "": usage/help message

examples::Vector{String} = String[]: usage examples

add_help::Bool = false: flag to automatically generate a help message

throw_on_exception = true: immediately throw on exception if true

SimpleArgParse.ArgumentValues — Type

ArgumentValues

Command-line argument values.

Fields

const args::ArgForms

value::Any

const type::Type = Any

const required::Bool = false

const positional::Bool = false

const description::String = ""

const validator::Union{AbstractValidator, Nothing} = nothing

SimpleArgParse.StrValidator — Type

StrValidator <: AbstractValidator

String validator.

Fields

upper_case::Bool = false: If true, input and pattern converted to uppercase, except for regex comparison

starts_with::Bool = false: If true, validate if one of the words in the patterns starts with input. Returns the whole matching word.

patterns::Vector{Union{AbstractString, Regex}}

Examples

julia> validate("foo", StrValidator(; upper_case=true, patterns=["foo", "bar"]))
(ok = true, v = "FOO")

julia> validate("foo", StrValidator(; patterns=[r"^fo[aoe]$"]))
(ok = true, v = "foo")

julia> validate("ye", StrValidator(; upper_case=true, starts_with=true, patterns=["yes", "no"]))
(ok = true, v = "YES")

Functions

SimpleArgParse._error — Method

_error(throw_on_exception, msg::AbstractString; excp=ArgumentError) → ::Exception

Depending on value of throw_on_exception, throw immediately, or return Exception to be processed downstream.

Function _error is internal.

SimpleArgParse.add_argument! — Function

add_argument!(parser::ArgumentParser, arg_short::String="", arg_long::String=""; kwargs...) → parser

Function add_argument! is exported

Arguments

parser::ArgumentParser: ArgumentParser object instance.

arg_short::String="": short argument flag.

arg_long::String="": long argument flag.

Keywords

type::Type=nothing: type, the argument value to be parsed/converted into.

default::Any=nothing

required::Bool=false

positional::Bool=false

description::String=nothing

validator::Union{AbstractValidator, Nothing}=nothing

SimpleArgParse.add_example! — Method

add_example!(parser::ArgumentParser, example::AbstractString) → parser

Function add_example! is exported.

SimpleArgParse.arg2strkey — Method

arg2strkey(arg::AbstractString) → ::SubString

Argument to argument-store string key conversion by removing hypenation from prefix.

Function arg2strkey is internal.

SimpleArgParse.args2vec — Method

args2vec(args::ArgForms) → ::Vector{String}

Extract struct members to vector of length 1 or 2.

Function args2vec is internal.

SimpleArgParse.colorize — Method

colorize(text; color, background, bright) → ::String

Colorize strings or backgrounds using ANSI codes and escape sequences.

Color	Example	Text	Background	Bright text	Bright background

Black	Black	30	40	90	100

Red	Red	31	41	91	101

Green	Green	32	42	92	102

Yellow	Yellow	33	43	93	103

Blue	Blue	34	44	94	104

Magenta	Magenta	35	45	95	105

Cyan	Cyan	36	46	96	106

White	White	37	47	97	107

Default		39	49	99	109

Arguments

text::String: the UTF-8/ASCII text to colorize.

Keywords

color::String="default": the standard ANSI name of the color.

background::Bool=false: flag to select foreground or background color.

bright::Bool=false: flag to select normal or bright text.

Function colorize is internal.

SimpleArgParse.colorprint — Function

colorprint(text, color="default", newline=true; background=false, bright=false) → nothing

Print colored text into stdout. For color table, see help to internal colorize function.

Function colorprint is exported.

SimpleArgParse.generate_usage! — Method

generate_usage!(parser::ArgumentParser) → ::String

Usage/help message generator.

Example of generated text

Usage: main.jl –input <PATH> [–verbose] [–problem] [–help]

A Julia script with command-line arguments.

Options: -i, –input <PATH> Path to the input file. -v, –verbose Enable verbose message output. -p, –problem Print the problem statement. -h, –help Print this help message.

Examples: $ julia main.jl –input dir/file.txt –verbose $ julia main.jl –help

SimpleArgParse.get_value — Method

get_value(parser, arg) → value::Any

Get argument value from parser.

Arguments

parser::ArgumentParser: ArgumentParser object instance.

arg::AbstractString="": argument name, e.g. "-f", "--foo".

Throws

Exception: depending on the value of parser.throw_on_exception::Bool, if the argument not found, the function will either throw imediately, or return e <: Exception to be processed downstream.

Function get_value is exported.

SimpleArgParse.help — Method

help(parser::ArgumentParser; color::AbstractString="default") → nothing

Print usage/help message.

SimpleArgParse.hyphenate — Method

hyphenate(argname::AbstractString) → ::String

Prepend hyphenation back onto argument after stripping it for the argument-store numkey.

Function hyphenate is internal.

SimpleArgParse.parse_args! — Method

parse_args!(parser::ArgumentParser; cli_args=nothing) → ::Union{ArgumentParser, Exception}

Parses arguments, validates them and stores the updated values in the parser.

Keywords

cli_args::Union{Vector{AbstractString}, Nothing}=nothing: if the cli_args not provided, parses the command line arguments ARGS. Otherwise accepts equivalent Vector of Strings, e.g. ["--foo", "FOO", "-i", "1"]

Throws

Exception: depending on the value of parser.throw_on_exception, in case of non-valid args vector, the function will either throw imediately, or return e <: Exception to be processed downstream.

Function parse_args! is exported.

SimpleArgParse.set_value! — Method

set_value!(parser::ArgumentParser, numkey::Integer, value::Any; throw_on_exception = true) → parser
set_value!(parser::ArgumentParser, argname::AbstractString, value::Any; throw_on_exception = true) → parser

Set/update value of argument, validating it, as specified by numkey or argname, in parser.

Throws

Exception: depending on the value of throw_on_exception kwarg, if the argument not found, the function will either throw imediately, or return e <: Exception to be processed downstream.

Function set_value! is exported.

SimpleArgParse.update_val! — Method

update_val!(parser::ArgumentParser, numkey::Integer, value) → parser

Function update_val! is internal

SimpleArgParse.validate — Method

validate(v::Any, ::Nothing) → (;ok=true, v)
validate(v::Nothing, ::Any) → (;ok=true, v)

If nothing is supplied instead of Validator, validation skipped. The same, if the value v to be validated is nothing.

SimpleArgParse.validate — Method

validate(v::Any, vl::AbstractValidator) → (;ok::Bool, v)

Validate input v against validator vl, and returns named tuple with validation result ok and (possibly canonicalized) input value v on success, or nothing on validation failure. For examples and specific information see documentation for the corresponding Validator, e.g. StrValidator or RealValidator.

Constants

SimpleArgParse.ANSICODES — Constant

Key-value store mapping from colors to ANSI codes.

Index

SimpleArgParse.ANSICODES

SimpleArgParse.AbstractValidator

SimpleArgParse.ArgForms

SimpleArgParse.ArgumentParser

SimpleArgParse.ArgumentValues

SimpleArgParse.StrValidator

SimpleArgParse._error

SimpleArgParse.add_argument!

SimpleArgParse.add_example!

SimpleArgParse.arg2strkey

SimpleArgParse.args2vec

SimpleArgParse.colorize

SimpleArgParse.colorprint

SimpleArgParse.generate_usage!

SimpleArgParse.get_value

SimpleArgParse.help

SimpleArgParse.hyphenate

SimpleArgParse.parse_args!

SimpleArgParse.set_value!

SimpleArgParse.update_val!

SimpleArgParse.validate

SimpleArgParse.validate