Types

SimpleArgParse2.AbstractValidator — Type

AbstractValidator

The supertype for validators. Type AbstractValidator is public, but not exported.

SimpleArgParse2.ArgForms — Type

ArgForms

Command-line arguments, short and long forms. Type ArgForms is exported.

Fields

short::String

long::String

SimpleArgParse2.ArgumentParser — Type

ArgumentParser <: AbstractArgumentParser

Command-line argument parser with numkey-value stores and attributes. Type AbstractArgumentParser is exported.

Fields

stores

kv_store::OrderedDict{UInt16,ArgumentValues} = OrderedDict(): numkey => value

arg_store::OrderedDict{String,UInt16} = OrderedDict(): numkey-value store: arg => numkey

lng::UInt16 = 0: counter of stored args

attributes

description::String = "": description

usage::String = "": The script name passed to Julia from the command line.

usage::String = "": usage/help message

examples::Vector{String} = String[]: usage examples

add_help::Bool = false: flag to automatically generate a help message

color::String = "default": output color - see also ANSICODES

SimpleArgParse2.ArgumentValues — Type

ArgumentValues

Command-line argument values. Type ArgumentValues is exported.

Fields

const args::ArgForms

value::Any

const type::Type = Any

const positional::Bool = false

const description::String = ""

const validator::Union{AbstractValidator, Nothing} = nothing

SimpleArgParse2.RealValidator — Type

RealValidator{T} <: AbstractValidator

Numbers validator type. If no include criteria specified, anything not excluded considered OK. The intervals are evaluated as closed a ≤ x ≤ b. Type RealValidator is exported.

Fields

excl_vals::Vector{T} = T[]: list of values to exclude

excl_ivls::Vector{Tuple{T, T}} = Tuple{T, T}[]: list of intervals to exclude

incl_vals::Vector{T} = T[]: list of accepted values

incl_ivls::Vector{Tuple{T, T}} = Tuple{T, T}[]: list of accepted intervals

Examples

julia> validate(1, RealValidator{Int}(;excl_vals=[1, 2], excl_ivls=[(10, 15), (20, 25)], incl_vals=[3, 4, 11], incl_ivls=[(100, 1000)]))
(ok = false, v = nothing)

julia> validate(150, RealValidator{Int}(;incl_ivls=[(100, 200)]))
(ok = true, v = 150)

julia> validate(50, RealValidator{Int}(;excl_ivls=[(100, 200)]))
(ok = true, v = 50)

SimpleArgParse2.StrValidator — Type

StrValidator <: AbstractValidator

String validator type. Type StrValidator is exported.

Fields

upper_case::Bool = false: If true, input and pattern converted to uppercase, except for regex comparison

starts_with::Bool = false: If true, validate if one of the words in the patterns starts with input. Returns the whole matching word.

patterns::Vector{Union{AbstractString, Regex}}

Examples

julia> validate("foo", StrValidator(; upper_case=true, patterns=["foo", "bar"]))
(ok = true, v = "FOO")

julia> validate("foo", StrValidator(; patterns=[r"^fo[aoe]$"]))
(ok = true, v = "foo")

julia> validate("ye", StrValidator(; upper_case=true, starts_with=true, patterns=["yes", "no"]))
(ok = true, v = "YES")

Functions

Exported functions

SimpleArgParse2.add_argument! — Function

add_argument!(parser::AbstractArgumentParser, arg_short::String="", arg_long::String=""; kwargs...) → Nothing

Arguments

parser::AbstractArgumentParser: AbstractArgumentParser object instance.

arg_short::String="": short argument flag.

arg_long::String="": long argument flag.

Keywords

type::Type=nothing: type, the argument value to be parsed/converted into.

default::Any=nothing

positional::Bool=false

description::String=nothing

validator::Union{AbstractValidator, Nothing}=nothing

Throws

Throws immediately in case of error, e.g. if a key already present.

Function add_argument! is exported

SimpleArgParse2.add_example! — Function

add_example!(parser::AbstractArgumentParser, example::AbstractString) → Nothing

Function add_example! is exported.

SimpleArgParse2.args_pairs — Function

args_pairs(parser::AbstractArgumentParser; excl::Union{Nothing, Vector{String}}=nothing) → ::Vector{Pair{Symbol, Any}}

Return vector of pairs argname => argvalue for all arguments except listed in excl. If argument has both short and long forms, the long one is used. Returned value can be e.g. passed as kwargs... to a function processing the parsed data, converted to a Dict or NamedTuple.

Function args_pairs is exported.

SimpleArgParse2.colorprint — Function

colorprint(text, color::AbstractString="default", newline=true; background=false, bright=false) → nothing
colorprint(text, parser::AbstractArgumentParser, newline=true; background=false, bright=false) → nothing

Print colored text into stdout. For color table, see help to internal colorize function. If second arg is an AbstractArgumentParser, uses color as defined within, if any, otherwise uses default.

Function colorprint is exported.

Base.haskey — Function

haskey(parser::AbstractArgumentParser, key::AbstractString) → ::Bool
haskey(parser::AbstractArgumentParser, key::Integer) → ::Bool

SimpleArgParse2.help — Function

help(parser::AbstractArgumentParser; color::Union{AbstractString, Nothing}) → nothing

Print usage/help message. Function help is exported.

SimpleArgParse2.parse_args! — Function

parse_args!(parser::AbstractArgumentParser; cli_args=nothing) →  ::Union{Nothing, Exception}

Parses arguments, validates them and stores the updated values in the parser.

Keywords

cli_args::Union{Vector{AbstractString}, Nothing}=nothing: if the cli_args not provided, parses the command line arguments ARGS. Otherwise accepts equivalent Vector of Strings, e.g. ["--foo", "FOO", "-i", "1"]

Throws

Exception: depending on the value of parser.interactive, in case of non-valid args vector, the function will either throw imediately, or return e <: Exception to be processed downstream.

Function parse_args! is exported.

Base.shell_split — Function

shell_split(s::AbstractString) → String[]

Split a string into a vector of args.

shell_split is in internal function of Base. It is re-exported.

Examples

julia> shell_split("--foo 3 -b bar")
4-element Vector{String}:
 "--foo"
 "3"
 "-b"
 "bar"

SimpleArgParse2.validate — Function

validate(v::Any, ::Nothing) → (;ok=true, v)
validate(v::Missing, ::Any) → (;ok=true, v)
validate(v::Any, vl::AbstractValidator) → (;ok::Bool, v)

Validate input v against validator vl, and returns named tuple with validation result ok and (possibly canonicalized) input value v on success, or nothing on validation failure. If nothing is supplied instead of Validator, validation skipped. The same, if the value v to be validated is nothing. For examples and specific information see documentation for the corresponding Validator, e.g. StrValidator or RealValidator. Function validate is exported.

Public functions

SimpleArgParse2.generate_usage! — Function

generate_usage!(parser::AbstractArgumentParser) → Nothing

Usage/help message generator. Function generate_usage! is public, not exported.

Example of generated text

Usage: main.jl –input <PATH> [–verbose] [–problem] [–help]

A Julia script with command-line arguments.

Options: -i, –input <PATH> Path to the input file. -v, –verbose Enable verbose message output. -p, –problem Print the problem statement. -h, –help Print this help message.

Examples: $ julia main.jl –input dir/file.txt –verbose $ julia main.jl –help

SimpleArgParse2.get_value — Function

get_value(parser, arg) → value::Any

Get argument value from parser.

Arguments

parser::AbstractArgumentParser: AbstractArgumentParser object instance.

arg::AbstractString="": argument name, e.g. "-f", "--foo".

Throws

Exception: depending on the value of throw_on_exception(parser), if the argument not found, the function will either throw imediately, or return e <: Exception to be processed downstream.

Function get_value is public, not exported.

SimpleArgParse2.getcolor — Function

getcolor(parser::AbstractArgumentParser, color=nothing)  → color::String

Returns color in case second arg is defined, otherwise the color defined in parser, or "default".

Function getcolor is public, not exported.

SimpleArgParse2.parse_arg — Function

parse_arg(t::Type, val_str::Union{AbstractString, Bool}, ::Union{Nothing, AbstractValidator}) → (; ok, v=parsed_value, msg=nothing)

Tries to parse val_str to type t. For your custom types or custom parsing, provide your own methods.

Function parse_arg is public, but not exported.

SimpleArgParse2.set_value! — Function

set_value!(parser::AbstractArgumentParser, numkey::Integer, value::Any) → ::Union{Nothing, Exception}
set_value!(parser::AbstractArgumentParser, argname::AbstractString, value::Any) → ::Union{Nothing, Exception}

Set/update value of argument, validating it, as specified by numkey or argname, in parser.

Throws

Exception: depending on the value of throw_on_exception , if the argument not found, the function will either throw imediately, or return e <: Exception to be processed downstream.

Function set_value! is public, not exported.

Internal functions

SimpleArgParse2._error — Method

_error(throw_on_exception, msg::AbstractString; excp=ArgumentError) → ::Exception

Depending on value of throw_on_exception, throw immediately, or return Exception to be processed downstream.

Function _error is internal.

SimpleArgParse2.arg2strkey — Method

arg2strkey(arg::AbstractString) → ::SubString

Argument to argument-store string key conversion by removing hypenation from prefix.

Function arg2strkey is internal.

SimpleArgParse2.args2vec — Method

args2vec(args::ArgForms) → ::Vector{String}

Extract struct members to vector of length 1 or 2.

Function args2vec is internal.

SimpleArgParse2.argument_usage — Method

argument_usage(v::ArgumentValues) → (; u=args_usage, o=options)

Function argument_usage is internal.

SimpleArgParse2.check_missing_input — Method

check_missing_input(parser::AbstractArgumentParser) → Union{Nothing, Exception}

Checks if all required arguments were supplied. Required is an argument without a default value.

Throws

Exception: depending on the value of throw_on_exception(parser), if an argument is missing, the function will either throw imediately, or return e <: Exception to be processed downstream.

Function check_missing_input is internal.

SimpleArgParse2.colorize — Method

colorize(text; color, background, bright) → ::String

Colorize strings or backgrounds using ANSI codes and escape sequences.

Color	Example	Text	Background	Bright text	Bright background

Black	Black	30	40	90	100

Red	Red	31	41	91	101

Green	Green	32	42	92	102

Yellow	Yellow	33	43	93	103

Blue	Blue	34	44	94	104

Magenta	Magenta	35	45	95	105

Cyan	Cyan	36	46	96	106

White	White	37	47	97	107

Default		39	49	99	109

Arguments

text::AbstractString: the UTF-8/ASCII text to colorize.

Keywords

color::AbstractString="default": the standard ANSI name of the color.

background::Bool=false: flag to select foreground or background color.

bright::Bool=false: flag to select normal or bright text.

Function colorize is internal.

SimpleArgParse2.hyphenate — Method

hyphenate(argname::AbstractString) → ::String

Prepend hyphenation back onto argument after stripping it for the argument-store numkey.

Function hyphenate is internal.

SimpleArgParse2.sort_args — Method

sort_args(parser::AbstractArgumentParser) → (;pos_args, keyed_args, all_args)

Function sort_args is internal.

SimpleArgParse2.update_val! — Method

update_val!(parser::AbstractArgumentParser, numkey::Integer, val_str::AbstractString) → ::Union{Nothing, Exception}

See also set_value!. Function update_val! is internal.

Constants

SimpleArgParse2.ANSICODES — Constant

Key-value store mapping from colors to ANSI codes. An internal constant.

Index

SimpleArgParse2.ANSICODES

SimpleArgParse2.AbstractValidator

SimpleArgParse2.ArgForms

SimpleArgParse2.ArgumentParser

SimpleArgParse2.ArgumentValues

SimpleArgParse2.RealValidator

SimpleArgParse2.StrValidator

Base.haskey

Base.shell_split

SimpleArgParse2._error

SimpleArgParse2.add_argument!

SimpleArgParse2.add_example!

SimpleArgParse2.arg2strkey

SimpleArgParse2.args2vec

SimpleArgParse2.args_pairs

SimpleArgParse2.argument_usage

SimpleArgParse2.check_missing_input

SimpleArgParse2.colorize

SimpleArgParse2.colorprint

SimpleArgParse2.generate_usage!

SimpleArgParse2.get_value

SimpleArgParse2.getcolor

SimpleArgParse2.help

SimpleArgParse2.hyphenate

SimpleArgParse2.parse_arg

SimpleArgParse2.parse_args!

SimpleArgParse2.set_value!

SimpleArgParse2.sort_args

SimpleArgParse2.update_val!

SimpleArgParse2.validate